refactor: split node.rs into a backend module
`node.rs` had grown to ~1900 lines holding six unrelated data source
implementations. Move each into its own file under `src/backend/`, leaving
`mod.rs` with the `Node` trait, its shared default fetching logic,
`Capabilities` and `NodeInfo`.

closes https://github.com/0xB10C/fork-observer/issues/147
